requirements.txt
streamlit>=1.30.0
pandas>=1.5.0
matplotlib>=3.6.0
requests>=2.28.0
openpyxl>=3.1.0
scikit-learn>=1.6.0
python-dotenv>=1.0.0
